Fix #7; set dtypes directly to avoid typecasting later

+++ radiant/cameras.py
@@ -12,11 +12,10 @@ class Camera:
         self.look_at(eye, target, up=up)
 
     def look_at(self, eye, target, up=None):
-        self.view = pyrr.Matrix44.look_at(eye, target, up or UP).astype('f4')
+        self.view = pyrr.Matrix44.look_at(eye, target, up or UP, dtype='f4')
 
 
 class PerspectiveCamera(Camera):
     def __init__(self, eye, target, up=None, fov=45.0, aspect=4.0/3.0, near=0.1, far=1000.0):
         super().__init__(eye, target, up=up)
-        self.projection = pyrr.Matrix44.perspective_projection(
-            fov, aspect, near, far, dtype='f4')
+        self.projection = pyrr.Matrix44.perspective_projection(fov, aspect, near, far, dtype='f4')
